Implement word prediction support using the new NLP core
The IPC communication causes significant lag atm.
